requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
coreapi==2.3.3
coreschema==0.0.4
Django==5.1.2
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
idna==3.10
itypes==1.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
mysqlclient==2.2.5
PyJWT==2.9.0
python-decouple==3.8
python-dotenv==1.0.1
requests==2.32.3
sqlparse==0.5.1
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
